+++ exeris-kernel-community/src/test/java/eu/exeris/kernel/community/http/CommunityHttpSecurityAdmissionIntegrationTest.java
@@ -435,7 +435,8 @@ class CommunityHttpSecurityAdmissionIntegrationTest {
                 HttpConfig.DEFAULT_MAX_REQUEST_BODY_BYTES,
                 false,
                 HttpVersion.HTTP_1_1,
-                "127.0.0.1" + ":" + port
+                "127.0.0.1" + ":" + port,
+                HttpConfig.DEFAULT_MAX_HEADER_BLOCK_SIZE
         );
     }